Test new deploy script

+++ scripts/deploy.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
-sudo apt-get install -qq ncftp
-
-ncftp -u "$USERNAME" -p "$PASSWORD" $HOST<<EOF
-rm -rf site/wwwroot
-mkdir site/wwwroot
+if  [[ $TRAVIS_PULL_REQUEST = "false" ]]
+then
+    ncftp -u "$USERNAME" -p "$PASSWORD" "$HOST"<<EOF
+    rm -rf site/wwwroot
+    mkdir site/wwwroot
+    quit
 EOF
 
-cd _site
-ncftpput -R -v -u "$USERNAME" -p "$PASSWORD" $HOST /site/wwwroot .
+    cd _site || exit
+    ncftpput -R -v -u "$USERNAME" -p "$PASSWORD" "$HOST" /site/wwwroot .
+fi
